requirements-dev.txt
asgiref==3.2.7
certifi==2020.4.5.1
chardet==3.0.4
dj-database-url==0.5.0
dj-static==0.0.6
Django==3.0.7
django-appconf==1.0.4
django-filter==2.2.0
django-filter-stubs==0.1.1
django-heroku-connect==0.19.3
django-stubs==1.5.0
djangorestframework==3.11.0
djangorestframework-stubs==1.2.0
idna==2.9
mypy==0.770
mypy-extensions==0.4.3
python-decouple==3.3
pytz==2020.1
requests==2.23.0
sqlparse==0.3.1
static3==0.7.0
typed-ast==1.4.1
typing-extensions==3.7.4.2
urllib3==1.25.9
validate-docbr==1.7.0
